refactor: auth 패키지 내부 분리
