+++ src/main/java/com/example/parking/external/config/argumentresolver/AuthArgumentResolver.java
@@ -1,7 +1,7 @@
 package com.example.parking.external.config.argumentresolver;
 
 import com.example.parking.auth.AuthService;
-import com.example.parking.auth.MemberSession;
+import com.example.parking.auth.session.MemberSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
